Modificaciones varias

+++ SistemaFacturacion_WebAssembly/Models/AuthTokenHandler.cs
@@ -11,6 +11,12 @@ namespace SistemaFacturacion_WebAssembly.Models
         private readonly ILocalStorageService _localStorageService;
         private readonly UsuarioEstadoService _userStateService;
 
+        // Lista de endpoints públicos
+        private readonly HashSet<string> _publicEndpoints = new HashSet<string>
+        {
+            "/api/Usuario/IniciarSesion", // Endpoint público
+        };
+
         public AuthTokenHandler(ILocalStorageService localStorageService, UsuarioEstadoService userStateService)
         {
             _localStorageService = localStorageService;
@@ -19,21 +25,30 @@ namespace SistemaFacturacion_WebAssembly.Models
 
         protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
-            var accessToken = _userStateService.Token ?? await _localStorageService.GetItemAsync<string>("accessToken");
+            // Verifica si la solicitud es para un endpoint público
+            var requestUri = request.RequestUri?.AbsolutePath;
+            var isPublicEndpoint = requestUri != null && _publicEndpoints.Contains(requestUri);
 
-            if (!string.IsNullOrEmpty(accessToken))
+
+            if (!isPublicEndpoint)
             {
-                var tokenHandler = new JwtSecurityTokenHandler();
-                var token = tokenHandler.ReadJwtToken(accessToken);
+                var accessToken = _userStateService.Token ?? await _localStorageService.GetItemAsync<string>("accessToken");
 
-                if (token.ValidTo < DateTime.UtcNow) // Token expirado
+                if (!string.IsNullOrEmpty(accessToken))
                 {
-                    await _localStorageService.RemoveItemAsync("accessToken");
-                    throw new UnauthorizedAccessException("El token ha expirado.");
-                }
+                    var tokenHandler = new JwtSecurityTokenHandler();
+                    var token = tokenHandler.ReadJwtToken(accessToken);
 
-                request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+                    if (token.ValidTo < DateTime.UtcNow) // Token expirado
+                    {
+                        await _localStorageService.RemoveItemAsync("accessToken");
+                        throw new UnauthorizedAccessException("El token ha expirado.");
+                    }
+
+                    request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+                }
             }
+                        
 
             var response = await base.SendAsync(request, cancellationToken);
